+++ src/resources/supplier.py
@@ -2,6 +2,7 @@ from flask_restful import Resource, reqparse
 from flask_jwt_extended import jwt_required
 from model.supplier import SupplierModel as Model
 from helpers.common import get_logger
+from helpers.user_role_management import role_has_permissions
 
 logger = get_logger(__name__)
 
@@ -32,6 +33,8 @@ class Supplier(Resource):
 
     @jwt_required()
     def get(self):
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         data = Supplier.get_parser.parse_args()
         obj_list = Model.find_by(**data)
         if obj_list and len(obj_list) > 0:
@@ -40,6 +43,8 @@ class Supplier(Resource):
 
     @jwt_required()
     def post(self):
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         data = Supplier.post_parser.parse_args()
         name = data.get("name")
 
@@ -60,6 +65,8 @@ class Supplier(Resource):
 
     @jwt_required()
     def put(self):
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         data = Supplier.post_parser.parse_args()
         name = data.get("name")
         
@@ -81,6 +88,8 @@ class Supplier(Resource):
 
     @jwt_required(fresh=True)
     def delete(self):
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         data = Supplier.get_parser.parse_args()
 
         obj_list = Model.find_by(**data)
@@ -94,11 +103,15 @@ class Suppliers(Resource):
 
     @jwt_required()
     def get(self):
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         return { 'Elements': [obj.json() for obj in Model.query.all()] }
     
     @jwt_required(fresh=True)
     def delete(self):
         # TODO: find if exists delete all
+        if not role_has_permissions(Model.__tablename__):
+            return { 'message': 'User does not have permissions to perform this request' }
         for obj in Model.query.all():
             obj.delete_from_db()
         return { 'message': 'Elements deleted from db' }
